move empty state to prevent edge case
